requirements.txt
Django==3.0.7
django-crispy-forms==1.9.1
mysqlclient==1.4.6
uWSGI==2.0.19.1